Update arch_test.h to support IAF testing
Adding logic into the  trap handler to support IAF testing. Approved solution in today's cvw-arch-verif meeting. Umer will add to this flag to generalize it and make it also applicable for PMP tests. It was also discussed in the meeting that using ra is fine since we need more registers than the 6 T defined at the top of the trap handler, plus ra is not being modified only read. 

Signed-off-by: Marina-Bellido <116534125+Marina-Bellido@users.noreply.github.com>