Conditon hypervisor state saving on test requirements
This change avoids unnecessary signature changes by ensuring
hypervisor-specific fields are only saved in the trap signature if a
test explicitly signals a dependency on the H-extension.  The trap
handler now checks for the presence of the `rvtest_vtrap_routine` macro.
If this macro is defined, the two additional hypervisor trap fields are
stored; otherwise, they are omitted, thus preserving the test signature
for non-H extension tests.