Correct the definitions of interrupt clearing tables based on supported modes
We have an interrupt clearing table with different definitions depending
on the modes supported on the machine.  Due to a typographical error, we
were mistakenly using the M/U table for M/S/U mode, and concatenating
the M/S/U table to the M/S/V/U table.  Appending the two tables caused
issues when trying to index the contiguously following exception handler
table with `rvtest_vtrap_routine` defined.