Fix returning to M from V=1
When the hypervisor is enabled, a test will typically pass
`-Drvtest_vtrap_routine=True`.  When testing `#if
(rvtest_vtrap_routine)`, this will evaluate to false, so we must instead
test with `#ifdef rvtest_vtrap_routine`. This fixes returning to M mode,
we now properly inspect the Machine Previous Virtualization Mode bit of
mstatus/mstatush to find the correct save area.