Fix GOTO_LOWER_MODE to VS/VU mode
When moving to a virtualized mode, we need to make sure we use the guest
save area.  The if statements incorrectly checks for a non-existent
Vmode, instead of the two virtualized modes Vsmode and VUmode.  This
fixes the condition of the original code.

This also fixes the code comment to reflect the code, which explicitly
sets V=0 for HS/HU modes, sets V=1 for VS/VU modes, and leaves V the
same for V/S modes.